init: lto: ensure initcall ordering
With LTO, the compiler doesn't necessarily obey the link order for
initcalls, and initcall variables need globally unique names to
avoid collisions at link time.

This change adds the __initcall_id() macro, which uses __COUNTER__
and __LINE__ to help ensure these variables have unique names, and
moves each variable to its own section when LTO is enabled, so the
correct order can be specified using a linker script.

The generate_initcall_ordering.pl script uses nm to find initcalls
from the object files passed to the linker, and generates a linker
script that specifies the intended order. With LTO, this script is
called in link-vmlinux.sh.

Signed-off-by: Sami Tolvanen <samitolvanen@google.com>