init/Kconfig: Unlock -O3 for all arches
Building a kernel with -O3 may help in hunting bugs like [1] and thus
using this switch should not be restricted to one specific arch only.

Thus lets expose it. If for some reasone we have to hide it, lets hide
it under EXPERT.

The commit is made against next-20191210 tag.

[1] https://lore.kernel.org/lkml/673b885183fb64f1cbb3ed2387524077@natalenko.name/

Signed-off-by: Oleksandr Natalenko <oleksandr@redhat.com>
Signed-off-by: Jesse Chan <jc@linux.com>
[nc: Replace 'enable' with 'Unlock' in commit message. This does not
     explicitly enable -O3, it merely allows it to be selected by
     architectures other than ARC. -O2 is still the default]
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>