Makefile: Allow LLVM_AR and LLVM_NM to be overridden
While llvm-ar and llvm-nm are the normal names for the binaries found in
most LLVM toolchains, there are some like the ones from apt.llvm.org
which have a version suffix, like llvm-ar-10 or llvm-nm-10. In this
scenario, the user has no way to override these variables and their
build will fail. Allow them to be overriden like CC or LD so that they
can use whatever version of the binary they want.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>