objtool: use sh_info to find the base for .rela sections
Use the section index in sh_info to find the base section for .rela
sections instead of relying on the section names to match. This is
needed because when initcall sections are merged in the correct order
with LTO, the corresponding .rela section isn't renamed, which causes
objtool to fail.

Signed-off-by: Sami Tolvanen <samitolvanen@google.com>