kernel/sys.c: Hijack 'uname -r' to show official WSL version
Some users reported to me that VS Code's Remote WSL plugin does not work
on my kernel. As it turns out, removing '-microsoft-standard' from the
version string breaks that extension, presumably because there is some
parsing of the version string under the hood (it does not appear the
source of that extension is public?)

One solution that I had was to just restore '-microsoft-standard' in the
version string. However, that will not restore the Remote WSL plugin to
its former glory. When lauching a file via 'code <file>', 'uname -r' is
parse with a specific regex that will not match this kernel's and
WSL_BUILD will not be set so we do not get some of the cool stuff like
git information.

As a result, this hack... it is so horrible but I cannot really think of
a better solution.

Link: https://github.com/microsoft/vscode/blob/763748e43cfa243d6e003a0f5fbfa1fcf6edec8c/resources/win32/bin/code.sh#L14
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>