x86/asm: mark __force_order weak to avoid modpost warnings
Clang may not optimize out references to __force_order in modules
when the integrated assembler is used, which trips modpost even
though the undefined symbol is harmless as it's not actually used.
This change marks __force_order weak to avoid the modpost warning
while still making the compiler think a variable is accessed.

Signed-off-by: Sami Tolvanen <samitolvanen@google.com>