mm: add generic __va_function and __pa_function
We use non-canonical CFI jump tables with CONFIG_CFI_CLANG, which
means the compiler replaces function address references with the
address of the function's CFI jump table entry. This results in
__pa_symbol(function), for example, returning the physical address of
the jump table entry, which can lead to address space confusion since
the jump table itself points to a virtual address.

This change adds generic definitions for __va_function and
__pa_function, which architectures that support CFI can override.

Signed-off-by: Sami Tolvanen <samitolvanen@google.com>