x86: add __va_function and __pa_function
We use non-canonical CFI jump tables with CONFIG_CFI_CLANG, which
means the compiler replaces function address references with the
address of the function's CFI jump table entry. This results in
__pa_symbol(function), for example, returning the physical address
of the jump table entry, which can lead to address space confusion
since the jump table itself points to a virtual address.

This change adds __va_function and __pa_function macros, which use
inline assembly to take the actual function address instead.

Signed-off-by: Sami Tolvanen <samitolvanen@google.com>